sample_semantic_kernel/requirements_SK.txt
# Python Requirements File
# Generated from main dependencies in pyproject.toml

# Core Framework
semantic-kernel>=1.33.0

# Azure AI Services
azure-ai-projects>=1.0.0b11
azure-ai-inference>=1.0.0b9
azure-identity>=1.21.0
azure-search-documents>=11.5.2
azure-ai-evaluation>=1.3.0

# Configuration & Environment
python-dotenv>=1.0.1
pydantic>=2.0.0
pyyaml>=6.0.0

# Web Framework & API
gradio>=5.20.1
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
psycopg>=3.2.6

# Testing
pytest>=8.3.5
pytest-mock>=3.14.0

# Development Tools
pyright>=1.1.400

# Jupyter & Notebooks
ipykernel>=6.29.5
jupyter>=1.0.0
nbformat
nbconvert

# Data Science & Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
pandas
pandas-stubs